fix: "db.sql BigDecimal to DOUBLE PRECISION"

+++ DB/db.sql
@@ -11,7 +11,7 @@ CREATE TABLE IF NOT EXISTS Devise (
 CREATE TABLE IF NOT EXISTS Transaction (
     id SERIAL PRIMARY KEY,
     label VARCHAR(255) NOT NULL,
-    sold numeric(10, 2) NOT NULL,
+    sold DOUBLE PRECISION NOT NULL,
     date_heure TIMESTAMP DEFAULT CURRENT_TIMESTAMP,
     type_transaction VARCHAR(10) CHECK (type_transaction IN ('debit', 'credit')) NOT NULL
 );
@@ -21,7 +21,7 @@ CREATE TABLE IF NOT EXISTS Transaction (
 CREATE TABLE IF NOT EXISTS Account (
     id SERIAL PRIMARY KEY,
     name VARCHAR(255) NOT NULL,
-    sold DECIMAL(10, 2) NOT NULL,
+    sold DOUBLE PRECISION NOT NULL,
     date_last_update TIMESTAMP DEFAULT CURRENT_TIMESTAMP,
     devise_id INT REFERENCES Devise(devise_id),
      type_compte VARCHAR(20) CHECK (type_compte IN ('Banque', 'Espèce', 'Mobile Money')) NOT NULL
